Modify stoneFlap servo control and add two modes of drive power
Needs to be tested! - there is a lot of commented-out code that can be cleaned up later.
The stoneFlap servo is now controlled with multiple rates of change by gamepad2's right stick. It can also be controlled by X and B buttons on gamepad2.
There are now high and low drive power levels controlled by gamepad1 buttons A and B respectively.

Signed-off-by: Richardio2002 <22622966+Richardio2002@users.noreply.github.com>